reddit-outreach-agent/requirements.txt
anthropic
langfuse
python-dotenv
mcp
streamlit
